Sanity checks on user name
Prevent empty user name and strip white spaces.
Limit length to 64, still allows for some creativity when choosing,
but limits annoyance to other users.
Prevent names from starting with /, wouldn't be much of an issue,
but could be annoying for other users wanting to address user with such a name.